+++ packages/core/src/cli/cli.test.ts
@@ -188,6 +188,20 @@ test('validate - requireReference - no reference, no upgradesFrom', async t => {
   t.true(error?.message.includes('does not specify what contract it upgrades from'), error?.message);
 });
 
+test('validate - requireReference and unsafeSkipStorageCheck', async t => {
+  const temp = await getTempDir(t);
+  const buildInfo = await artifacts.getBuildInfo(`contracts/test/cli/Validate.sol:StorageV1`);
+  await fs.writeFile(path.join(temp, 'validate.json'), JSON.stringify(buildInfo));
+
+  const error = await t.throwsAsync(
+    execAsync(`${CLI} validate ${temp} --contract StorageV1 --requireReference --unsafeSkipStorageCheck`),
+  );
+  t.true(
+    error?.message.includes('The requireReference and unsafeSkipStorageCheck options cannot be used at the same time.'),
+    error?.message,
+  );
+});
+
 test('validate - requireReference - no reference, has upgradesFrom - safe', async t => {
   const temp = await getTempDir(t);
   const buildInfo = await artifacts.getBuildInfo(`contracts/test/cli/Validate.sol:BecomesSafe`);